requirements.txt
# Core Dependencies
flask==2.3.3
python-dotenv==1.0.0

# LangChain and Related
langchain==0.1.0
langchain_community==0.0.10
langchain_openai==0.0.3
langchain_experimental==0.0.42
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==3.17.1

# Development
-e .